Refactor and clean up the PS/2 controller driver
