+++ kernel/ps2/Cargo.toml
@@ -1,11 +1,12 @@
 [package]
 name = "ps2"
 version = "0.1.0"
-description = "Drivers supporting the PS/2 interface, for legacy keyboard/mouse devices"
+description = "Drivers for the PS/2 interface, supporting legacy keyboard and mouse"
 authors = ["Bowen Liu <liubowenbob@hotmail.com>, Hecatia Elegua"]
 edition = "2021"
 
 [dependencies]
+log = "0.4.8"
 spin = "0.9.4"
 modular-bitfield = "0.11.2"
 
@@ -13,11 +14,14 @@ modular-bitfield = "0.11.2"
 version = "0.5.7"
 default-features = false
 
-[dependencies.log]
-version = "0.4.8"
-
 [dependencies.port_io]
 path = "../../libs/port_io"
+
+[dependencies.acpi]
+path = "../acpi/"
+
+[dependencies.fadt]
+path = "../acpi/fadt"
 
 [lib]
 crate-type = ["rlib"]
