+++ kernel/ps2/src/lib.rs
@@ -1,111 +1,121 @@
+//! Driver for the PS/2 interface, which supports legacy keyboard and mouse devices.
+//!
+//! The [`init()`] function sets up the single system-wide [`PS2Controller`],
+//! which can then be used to obtain a reference to a [`PS2Keyboard`] or [`PS2Mouse`].
+
 #![no_std]
 
-use log::warn;
+use log::{debug, warn};
+use modular_bitfield::{specifiers::{B1, B4, B8}, bitfield, BitfieldSpecifier};
 use num_enum::TryFromPrimitive;
 use port_io::Port;
-use spin::Mutex;
-use modular_bitfield::{specifiers::{B1, B4, B8}, bitfield, BitfieldSpecifier};
+use spin::{Mutex, Once};
+use fadt::Fadt;
 
 use HostToControllerCommand::*;
 use HostToKeyboardCommand::*;
 use HostToMouseCommand::*;
 use DeviceToHostResponse::*;
 
-/// Port used by PS/2 Controller and devices
-const PS2_DATA_PORT: u16 = 0x60;
-/// Port used to send commands to and receive status from the PS/2 Controller
-const PS2_COMMAND_AND_STATUS_PORT: u16 = 0x64;
 
-// https://wiki.osdev.org/%228042%22_PS/2_Controller#PS.2F2_Controller_Commands
-// quite a few of these are commented out because they're either unused, deprecated or non-standard.
-/// Command types which can be sent to the PS/2 Controller
-/// 
-/// Naming perspective: Output means "out of the device into the host".
-/// e.g. [WriteByteToPort2InputBuffer] means "write byte out of the host into port 2"
-pub enum HostToControllerCommand {
-    /// returns [ControllerConfigurationByte]
-    ReadFromInternalRAMByte0 = 0x20,
+/// The system-wide singleton PS/2 Controller.
+static PS2_CONTROLLER: Once<PS2Controller> = Once::new();
+
+
+/// Initializes the PS/2 controller, and the first and second PS/2 ports (if they exist).
+///
+/// We roughly follow the procedure from the [OS dev wiki].
+///
+/// [OS dev wiki]: https://wiki.osdev.org/%228042%22_PS/2_Controller#Initialising_the_PS.2F2_Controller
+pub fn init() -> Result<&'static PS2Controller, &'static str> {
+    if PS2_CONTROLLER.is_completed() {
+        return Err("The PS/2 Controller has already been initialized once.");
+    }
+
+    // Step 1: Initialise USB Controllers
+    // no USB support yet
+
+    // Step 2: Determine if the PS/2 Controller Exists.
+    // If ACPI, and therefore the FADT, is unsupported, the PS/2 controller is assumed to exist.
+    let acpi_tables = acpi::get_acpi_tables().lock();
+    if let Some(fadt) = Fadt::get(&acpi_tables) {
+        // If earlier than ACPI v2, the PS/2 controller is assumed to exist
+        if fadt.header.revision > 1 {
+            let has_controller = fadt.iapc_boot_architecture_flags & 0b10 == 0b10;
+            if !has_controller {
+                return Err("no PS/2 Controller present");
+            }
+        }
+    }
+
+    // Here: the PS/2 Controller exists, so create the object representing it.
+    let controller = PS2Controller {
+        data_port: Mutex::new(Port::new(PS2Controller::PS2_DATA_PORT)),
+        command_and_status_port: Mutex::new(Port::new(PS2Controller::PS2_COMMAND_AND_STATUS_PORT))
+    };
+
+    // Step 3: Disable Devices
+    controller.write_command(DisablePort1);
+    controller.write_command(DisablePort2);
+
+    // Step 4: Flush The Output Buffer
+    controller.flush_output_buffer();
+
+    // Step 5: Set the Controller Configuration Byte
+    let mut config = controller.read_config();
+    config.set_port1_interrupt_enabled(false);
+    config.set_port2_interrupt_enabled(false);
+    config.set_port1_scancode_translation_enabled(false);
+    let has_mouse = config.port2_clock_disabled();
+    controller.write_config(config);
+
+    // Step 6: Perform Controller Self Test
+    controller.test()?;
+    debug!("passed PS/2 controller test");
+    // as this can reset the controller on some hardware, we restore the config
+    controller.write_config(config);
+
+    // Step 7: Determine If There Are 2 Channels
+    // not needed, already done above
+
+    // Step 8: Perform Interface Tests
+    let port_1_works = controller.test_port1().is_ok();
+    let port_2_works = if has_mouse {
+        controller.test_port2().is_ok()
+    } else {
+        warn!("PS/2 controller is not dual channel, so no mouse available");
+        false
+    };
+    if !port_1_works && !port_2_works {
+        return Err("failed both PS/2 port tests, terminating PS/2 controller driver");
+    }
+
+    // Step 9: Enable Devices
+    if port_1_works {
+        debug!("passed PS/2 port 1 test");
+        controller.write_command(EnablePort1);
+        config.set_port1_clock_disabled(false);
+        config.set_port1_interrupt_enabled(true);
+    }
+    if port_2_works {
+        debug!("passed PS/2 port 2 test");
+        controller.write_command(EnablePort2);
+        config.set_port2_clock_disabled(false);
+        config.set_port2_interrupt_enabled(true);
+    }
+    controller.write_config(config);
+
+    // Step 10: Reset Devices
+    controller.keyboard_ref().reset()?;
+    controller.mouse_ref().reset()?;
+
+    debug!("Final PS/2 {:?}", controller.read_config());
     
-    // return is non-standard
-    // /// 0x21-0x3F; N is the command byte & 0x1F
-    // ReadFromInternalRAMByteN = 0x21,
-
-    /// sets [ControllerConfigurationByte]
-    WriteToInternalRAMByte0 = 0x60,
-
-    // usage is non-standard
-    // /// 0x61-0x7F; N is the command byte & 0x1F
-    // WriteToInternalRAMByteN = 0x61,
-
-    // these are all deprecated 
-    // PasswordInstalledTest = 0xA4,
-    // LoadSecurity = 0xA5,
-    // EnableSecurity = 0xA6,
-    
-    /// sets [ControllerConfigurationByte] `port2_clock_disabled`
-    /// 
-    /// Note: only if 2 PS/2 ports supported
-    DisablePort2 = 0xA7,
-    /// clears [ControllerConfigurationByte] `port2_clock_disabled`
-    /// 
-    /// Note: only if 2 PS/2 ports supported
-    EnablePort2 = 0xA8,
-    /// returns [PortTestResult]
-    /// 
-    /// Note: only if 2 PS/2 ports supported
-    TestPort2 = 0xA9,
-    /// see [PS2Controller::test]
-    TestController = 0xAA,
-    /// returns [PortTestResult]
-    TestPort1 = 0xAB,
-
-    // unused
-    // /// read all bytes of internal RAM
-    // DiagnosticDump = 0xAC,
-
-    /// sets [ControllerConfigurationByte] `port1_clock_disabled`
-    DisablePort1 = 0xAD,
-    /// clears [ControllerConfigurationByte] `port1_clock_disabled`
-    EnablePort1 = 0xAE,
-
-    // return is non-standard
-    // ReadControllerInputPort = 0xC0,
-
-    // both unused
-    // CopyBits0to3ofInputPortToStatusBits4to7 = 0xC1,
-    // CopyBits4to7ofInputPortToStatusBits4to7 = 0xC2,
-
-    // both unused and therefore not implemented, https://wiki.osdev.org/%228042%22_PS/2_Controller#PS.2F2_Controller_Output_Port
-    // /// returns [ControllerOutputPort]
-    // ReadControllerOutputPort = 0xD0,
-    // /// sets [ControllerOutputPort]
-    // ///
-    // /// Note: Check if the output buffer is empty first
-    // WriteByteToControllerOutputPort = 0xD1,
-
-    // both unused
-    // /// makes it look like the byte written was received from the first PS/2 port
-    // /// 
-    // /// Note: only if 2 PS/2 ports supported
-    // WriteByteToPort1OutputBuffer = 0xD2,
-    // /// makes it look like the byte written was received from the second PS/2 port
-    // /// 
-    // /// Note: only if 2 PS/2 ports supported
-    // WriteByteToPort2OutputBuffer = 0xD3,
-
-    /// sends next byte to the second PS/2 port
-    /// 
-    /// Note: only if 2 PS/2 ports supported
-    WriteByteToPort2InputBuffer = 0xD4,
-
-    // both unused
-    // ReadTestInputs = 0xE0,
-    // /// 0xF0-0xFF; Bits 0 to 3 correspond to 4 different output lines and are used as a mask:
-    // /// 0 = pulse line, 1 = don't pulse line; Bit 0 corresponds to the "reset" line.
-    // /// The other output lines don't have a standard/defined purpose.
-    // PulseOutputLineLowFor6ms = 0xF0,
+    Ok(PS2_CONTROLLER.call_once(|| controller))
 }
 
+/// An initialized PS/2 controller that can be used to communicate with
+/// legacy PS/2 keyboard and mouse devices.
 pub struct PS2Controller {
     /// at 0x60
     data_port: Mutex<Port<u8>>,
@@ -114,55 +124,49 @@ pub struct PS2Controller {
 }
 
 impl PS2Controller {
-    /// Constructs a new PS2 Controller.
-    ///
-    /// # Safety
-    /// 
-    /// Must only be called once.
-    pub const unsafe fn new() -> Self {
-        Self {
-            data_port: Mutex::new(Port::new(PS2_DATA_PORT)),
-            command_and_status_port: Mutex::new(Port::new(PS2_COMMAND_AND_STATUS_PORT))
-        }
-    }
+    /// Port used to transmit data to and from the PS/2 Controller and its devices.
+    const PS2_DATA_PORT: u16 = 0x60;
+    /// Port used to send commands to and receive status from the PS/2 Controller.
+    const PS2_COMMAND_AND_STATUS_PORT: u16 = 0x64;
 
-    /// Get a handle to the keyboard.
-    /// 
+    /// Returns a reference to the keyboard attached to this PS/2 controller.
+    ///
     /// The keyboard only uses the data port.
-    pub fn keyboard_handle(&self) -> PS2Keyboard {
+    pub fn keyboard_ref(&self) -> PS2Keyboard {
         PS2Keyboard::new(self)
     }
 
-    /// Get a handle to the keyboard.
-    /// 
+    /// Returns a reference to the mouse attached to this PS/2 controller.
+    ///
     /// The mouse uses both the data port and the command/status port.
-    pub fn mouse_handle(&self) -> PS2Mouse {
+    pub fn mouse_ref(&self) -> PS2Mouse {
         PS2Mouse::new(self)
     }
 
-    /// Write a command to the PS/2 command port/register
-    /// 
-    /// Note: Devices attached to the controller should be disabled
-    /// before sending commands that return data, otherwise the output buffer could get overwritten
+    /// Writes a command to the PS/2 command port.
+    ///
+    /// ## Usage Note
+    /// Devices attached to the controller should be disabled *before* writing
+    /// a command that returns data, otherwise the output buffer may get overwritten.
     pub fn write_command(&self, value: HostToControllerCommand) {
         unsafe {
             self.command_and_status_port.lock().write(value as u8);
         }
     }
 
-    /// read the config of the PS/2 port
+    /// Reads and returns the current config value from the PS/2 controller.
     pub fn read_config(&self) -> ControllerConfigurationByte {
         self.write_command(ReadFromInternalRAMByte0);
         ControllerConfigurationByte::from_bytes([self.read_data()])
     }
     
-    /// write the new config to the PS/2 command port (0x64)
+    /// Writes the given config `value` to the PS/2 controller.
     pub fn write_config(&self, value: ControllerConfigurationByte) {
         self.write_command(WriteToInternalRAMByte0);
         self.write_data(WritableData::Configuration(value));
     }
 
-    /// Read the PS/2 status port/register
+    /// Reads the PS/2 status port.
     pub fn status_register(&self) -> ControllerToHostStatus {
         ControllerToHostStatus::from_bytes([self.command_and_status_port.lock().read()])
     }
@@ -174,7 +178,7 @@ impl PS2Controller {
 
     pub fn test(&self) -> Result<(), &'static str> {
         self.write_command(TestController);
-        self.read_controller_test_result() //TODO: bad naming
+        self.read_controller_test_result()
     }
 
     pub fn test_port1(&self) -> Result<(), &'static str> {
@@ -197,10 +201,10 @@ impl PS2Controller {
         use PortTestResult::*;
         match self.read_port_test_result()? {
             Passed => Ok(()),
-            ClockLineStuckLow => Err("clock line stuck low"),
+            ClockLineStuckLow  => Err("clock line stuck low"),
             ClockLineStuckHigh => Err("clock line stuck high"),
-            DataLineStuckLow => Err("data line stuck low"),
-            DataLineStuckHigh => Err("data line stuck high"),
+            DataLineStuckLow   => Err("data line stuck low"),
+            DataLineStuckHigh  => Err("data line stuck high"),
         }
     }
 
@@ -255,7 +259,7 @@ impl PS2Controller {
     }
 
     /// Read data from the PS/2 data port.
-    /// 
+    ///
     /// Note: this can be called directly when within a PS/2 device's interrupt handler.
     /// If polling is used (e.g. in a device's init function), [ControllerToHostStatus] `output_buffer_full` needs to be checked first.
     fn read_data(&self) -> u8 {
@@ -263,7 +267,7 @@ impl PS2Controller {
     }
 
     /// Write data to the PS/2 data port.
-    /// 
+    ///
     /// Note: when writing to the PS/2 controller, it might be necessary to check for
     /// [ControllerToHostStatus] `!input_buffer_full` first.
     fn write_data(&self, value: WritableData) {
@@ -271,19 +275,21 @@ impl PS2Controller {
     }
 }
 
+/// A reference to the PS/2 Mouse connected to the PS/2 Controller.
 pub struct PS2Mouse<'c> {
     controller: &'c PS2Controller,
     id: MouseId
 }
 
 impl<'c> PS2Mouse<'c> {
-    /// Give a controller reference to the mouse.
+    /// Create a new reference to the mouse connected to the given PS2 controller.
+    ///
     /// The default mouse id of PS/2 is zero.
     pub fn new(controller: &'c PS2Controller) -> Self {
         Self { controller, id: MouseId::Zero }
     }
 
-    /// reset the mouse
+    /// Reset the mouse.
     pub fn reset(&self) -> Result<(), &'static str> {
         self.command_to_mouse(HostToMouseCommandOrData::MouseCommand(Reset))?; 
         if let Ok(SelfTestPassed) = self.controller.read_data().try_into() {
@@ -298,7 +304,7 @@ impl<'c> PS2Mouse<'c> {
     fn command_to_mouse(&self, value: HostToMouseCommandOrData) -> Result<(), &'static str> {
         self.controller.write_command(WriteByteToPort2InputBuffer);
         const RETRIES: u8 = 3;
-        for _ in 0..=RETRIES {
+        for _ in 0 ..= RETRIES {
             return match self.controller.polling_send_receive(HostToDevice::Mouse(value.clone()))? {
                 Acknowledge => Ok(()),
                 ResendCommand => continue,
@@ -309,7 +315,7 @@ impl<'c> PS2Mouse<'c> {
     }
 
     /// set PS/2 mouse's sampling rate
-    fn set_mouse_sampling_rate(&self, value: SampleRate) -> Result<(), &'static str> {
+    fn set_mouse_sampling_rate(&self, value: MouseSampleRate) -> Result<(), &'static str> {
         self.command_to_mouse(HostToMouseCommandOrData::MouseCommand(SampleRate))
             .and_then(|_| self.command_to_mouse(HostToMouseCommandOrData::SampleRate(value)))
             .map_err(|_| "failed to set the mouse sampling rate")
@@ -319,7 +325,7 @@ impl<'c> PS2Mouse<'c> {
     pub fn set_mouse_id(&mut self, id: MouseId) -> Result<(), &'static str> {
         self.disable_mouse_packet_streaming()?;
 
-        use crate::SampleRate::*;
+        use crate::MouseSampleRate::*;
         match id {
             MouseId::Zero => { /* do nothing */ }
             MouseId::Three => {
@@ -345,7 +351,10 @@ impl<'c> PS2Mouse<'c> {
         self.disable_mouse_packet_streaming()?;
 
         self.command_to_mouse(HostToMouseCommandOrData::MouseCommand(GetDeviceID))?;
-        let id = self.controller.read_data().try_into().map_err(|_| "failed to get mouse id: bad response")?;
+        let id = self.controller
+            .read_data()
+            .try_into()
+            .map_err(|_| "failed to get mouse id: bad response")?;
 
         self.enable_mouse_packet_streaming()?;
         Ok(id)
@@ -402,22 +411,25 @@ impl<'c> PS2Mouse<'c> {
         }
     }
 
-    /// read the PS2Controller's status register to find out if the mouse output buffer is full, i.e. can be read
+    /// Returns `true` if there is content in the PS/2 Mouse's output buffer
+    /// that can be read from.
     pub fn is_output_buffer_full(&self) -> bool {
         self.controller.status_register().mouse_output_buffer_full()
     }
 }
 
+/// A reference to the PS/2 Keyboard connected to the PS/2 Controller.
 pub struct PS2Keyboard<'c> {
     controller: &'c PS2Controller,
 }
 
 impl<'c> PS2Keyboard<'c> {
+    /// Create a new reference to the keyboard connected to the given PS2 controller.
     pub fn new(controller: &'c PS2Controller) -> Self {
         Self { controller }
     }
     
-    /// reset the keyboard
+    /// Reset the keyboard.
     pub fn reset(&self) -> Result<(), &'static str> {
         self.command_to_keyboard(HostToKeyboardCommandOrData::KeyboardCommand(ResetAndStartSelfTest))?; 
         if let Ok(SelfTestPassed) = self.controller.read_data().try_into() {
@@ -439,22 +451,22 @@ impl<'c> PS2Keyboard<'c> {
         Err("keyboard doesn't support the command or there has been a hardware failure")
     }
 
-    /// set LED status of the keyboard
+    /// Set the status LEDs on the keyboard.
     pub fn set_keyboard_led(&self, value: LEDState) -> Result<(), &'static str> {
         self.command_to_keyboard(HostToKeyboardCommandOrData::KeyboardCommand(SetLEDStatus))
             .and_then(|_| self.command_to_keyboard(HostToKeyboardCommandOrData::LEDState(value)))
             .map_err(|_| "failed to set the keyboard led")
     }
 
-    /// set the scancode set of the keyboard
+    /// Set the active scancode set currently used by the keyboard.
     pub fn keyboard_scancode_set(&self, value: ScancodeSet) -> Result<(), &'static str> {
         self.command_to_keyboard(HostToKeyboardCommandOrData::KeyboardCommand(SetScancodeSet))
             .and_then(|_| self.command_to_keyboard(HostToKeyboardCommandOrData::ScancodeSet(value)))
             .map_err(|_| "failed to set the keyboard scancode set")
     }
 
-    /// detect the [KeyboardType]
-    /// 
+    /// Detect the [KeyboardType].
+    ///
     /// Note:
     /// On the identify command, [KeyboardType::AncientATKeyboard] usually returns no bytes at all, but `PS2Controller::read_data` always returns a byte.
     /// This means `PS2Controller::read_data` would return 0x00 here, even though the value is already reserved for the device type "Standard PS/2 mouse".
@@ -479,27 +491,122 @@ impl<'c> PS2Keyboard<'c> {
         keyboard_type
     }
 
-    /// reads a Scancode (for now an untyped u8)
+    /// Reads a raw scancode value from the PS/2 controller's data port.
     pub fn read_scancode(&self) -> u8 {
         self.controller.read_data()
     }
+}
+
+
+// https://wiki.osdev.org/%228042%22_PS/2_Controller#PS.2F2_Controller_Commands
+// quite a few of these are commented out because they're either unused, deprecated or non-standard.
+/// Commands that can be sent to the PS/2 Controller.
+///
+/// The naming convention is device-centric:
+/// * "Output" means sending data "out of the device to the host (CPU)".
+/// * "Input" means sending data "out of the host (CPU) into the device".
+///   * Example: [WriteByteToPort2InputBuffer] instructs the host (CPU) to write a byte
+///     into port 2 of the device, so the device receives that byte as input.
+pub enum HostToControllerCommand {
+    /// returns [ControllerConfigurationByte]
+    ReadFromInternalRAMByte0 = 0x20,
+    
+    // return is non-standard
+    // /// 0x21-0x3F; N is the command byte & 0x1F
+    // ReadFromInternalRAMByteN = 0x21,
+
+    /// sets [ControllerConfigurationByte]
+    WriteToInternalRAMByte0 = 0x60,
+
+    // usage is non-standard
+    // /// 0x61-0x7F; N is the command byte & 0x1F
+    // WriteToInternalRAMByteN = 0x61,
+
+    // these are all deprecated 
+    // PasswordInstalledTest = 0xA4,
+    // LoadSecurity = 0xA5,
+    // EnableSecurity = 0xA6,
+    
+    /// sets [ControllerConfigurationByte] `port2_clock_disabled`
+    ///
+    /// Note: only if 2 PS/2 ports supported
+    DisablePort2 = 0xA7,
+    /// clears [ControllerConfigurationByte] `port2_clock_disabled`
+    ///
+    /// Note: only if 2 PS/2 ports supported
+    EnablePort2 = 0xA8,
+    /// returns [PortTestResult]
+    ///
+    /// Note: only if 2 PS/2 ports supported
+    TestPort2 = 0xA9,
+    /// see [PS2Controller::test]
+    TestController = 0xAA,
+    /// returns [PortTestResult]
+    TestPort1 = 0xAB,
+
+    // unused
+    // /// read all bytes of internal RAM
+    // DiagnosticDump = 0xAC,
+
+    /// sets [ControllerConfigurationByte] `port1_clock_disabled`
+    DisablePort1 = 0xAD,
+    /// clears [ControllerConfigurationByte] `port1_clock_disabled`
+    EnablePort1 = 0xAE,
+
+    // return is non-standard
+    // ReadControllerInputPort = 0xC0,
+
+    // both unused
+    // CopyBits0to3ofInputPortToStatusBits4to7 = 0xC1,
+    // CopyBits4to7ofInputPortToStatusBits4to7 = 0xC2,
+
+    // both unused and therefore not implemented, https://wiki.osdev.org/%228042%22_PS/2_Controller#PS.2F2_Controller_Output_Port
+    // /// returns [ControllerOutputPort]
+    // ReadControllerOutputPort = 0xD0,
+    // /// sets [ControllerOutputPort]
+    // ///
+    // /// Note: Check if the output buffer is empty first
+    // WriteByteToControllerOutputPort = 0xD1,
+
+    // both unused
+    // /// makes it look like the byte written was received from the first PS/2 port
+    // ///
+    // /// Note: only if 2 PS/2 ports supported
+    // WriteByteToPort1OutputBuffer = 0xD2,
+    // /// makes it look like the byte written was received from the second PS/2 port
+    // ///
+    // /// Note: only if 2 PS/2 ports supported
+    // WriteByteToPort2OutputBuffer = 0xD3,
+
+    /// sends next byte to the second PS/2 port
+    ///
+    /// Note: only if 2 PS/2 ports supported
+    WriteByteToPort2InputBuffer = 0xD4,
+
+    // both unused
+    // ReadTestInputs = 0xE0,
+    // /// 0xF0-0xFF; Bits 0 to 3 correspond to 4 different output lines and are used as a mask:
+    // /// 0 = pulse line, 1 = don't pulse line; Bit 0 corresponds to the "reset" line.
+    // /// The other output lines don't have a standard/defined purpose.
+    // PulseOutputLineLowFor6ms = 0xF0,
 }
 
 // see https://wiki.osdev.org/%228042%22_PS/2_Controller#Status_Register
 // and https://users.utcluj.ro/~baruch/sie/labor/PS2/PS-2_Keyboard_Interface.htm
 #[bitfield(bits = 8)]
 pub struct ControllerToHostStatus {
-    /// When using [Polling](https://wiki.osdev.org/%228042%22_PS/2_Controller#Polling), must be `true` before attempting to read data from [PS2_DATA_PORT].
+    /// When using [Polling](https://wiki.osdev.org/%228042%22_PS/2_Controller#Polling), must be `true` before attempting to read data from PS/2 data port.
     /// When using [Interrupts](https://wiki.osdev.org/%228042%22_PS/2_Controller#Interrupts), guaranteed to be `true`, because an interrupt only happens if the buffer is full
     /// Alternative name: output_register_full, which might make more sense because it only ever fits 1 byte. "Buffer" sounds like multiple.
     output_buffer_full: bool,
-    /// Must be `false` before attempting to write data to [PS2_DATA_PORT] or [PS2_COMMAND_AND_STATUS_PORT] for 8042 keyboard controller
+    /// Must be `false` before attempting to write data to PS/2 data port or PS/2 command/status port for 8042 keyboard controller
     /// Alternative name: input_register_full, which might make more sense because it only ever fits 1 byte. "Buffer" sounds like multiple.
     input_buffer_full: bool,
     /// Cleared on reset; set when the system passed Power-on self-test
     #[allow(dead_code)]
     system_passed_self_test: bool,
-    /// Input buffer should be written to: 0 - [PS2_DATA_PORT], 1 - [PS2_COMMAND_AND_STATUS_PORT]
+    /// If `0`, the input buffer should be written to the PS/2 data port.
+    /// If `1`, the input buffer should be written to the PS/2 command/status port.
     #[allow(dead_code)]
     input_buffer_is_command: B1,
     /// Whether or not communication is inhibited (via switch) / Unknown (chipset specific) / (more likely unused on modern systems)
@@ -522,7 +629,7 @@ pub struct ControllerToHostStatus {
     parity_error: bool,
 }
 
-/// All types of data writable to the [PS2_DATA_PORT].
+/// Kinds of data that can be written to the PS/2 data port.
 enum WritableData {
     Configuration(ControllerConfigurationByte),
     HostToDevice(HostToDevice)
@@ -550,15 +657,19 @@ impl From<WritableData> for u8 {
     }
 }
 
-// wiki.osdev.org/%228042%22_PS/2_Controller#PS.2F2_Controller_Configuration_Byte
-/// Used for [HostToControllerCommand::ReadFromInternalRAMByte0] and [HostToControllerCommand::WriteToInternalRAMByte0]
+/// Configuration values used to enable or disable features of the PS/2 controller.
+///
+/// This type is used with [HostToControllerCommand::ReadFromInternalRAMByte0]
+/// and [HostToControllerCommand::WriteToInternalRAMByte0].
+///
+/// [More info](wiki.osdev.org/%228042%22_PS/2_Controller#PS.2F2_Controller_Configuration_Byte)
 #[bitfield(bits = 8)]
 #[derive(Debug, Clone, Copy)]
 pub struct ControllerConfigurationByte {
     /// interrupt on [ControllerToHostStatus] `output_buffer_full`
     pub port1_interrupt_enabled: bool,
     /// interrupt on [ControllerToHostStatus] `mouse_output_buffer_full`
-    /// 
+    ///
     /// Note: only if 2 PS/2 ports supported
     pub port2_interrupt_enabled: bool,
     /// Cleared on reset; set when the system passed Power-on self-test
@@ -570,7 +681,7 @@ pub struct ControllerConfigurationByte {
     /// disables the keyboard
     pub port1_clock_disabled: bool,
     /// disables the auxilary device (mouse)
-    /// 
+    ///
     /// Note: only if 2 PS/2 ports supported
     pub port2_clock_disabled: bool,
     /// whether IBM scancode translation is enabled (0=AT, 1=PC)
@@ -670,7 +781,7 @@ pub enum ScancodeSet {
 #[derive(Clone)]
 pub enum HostToMouseCommandOrData {
     MouseCommand(HostToMouseCommand),
-    SampleRate(SampleRate),
+    SampleRate(MouseSampleRate),
     MouseResolution(MouseResolution),
 }
 
@@ -721,7 +832,7 @@ pub struct MousePacketGeneric {
 
 impl MousePacketGeneric {
     /// `x_1st_to_8th_bit` and `x_9th_bit` should not be accessed directly, because they're part of one signed 9-bit number
-    /// 
+    ///
     /// Note: This handles overflow as well, although it might not happen on newer systems <https://forum.osdev.org/viewtopic.php?f=1&t=31176>
     pub fn x_movement(&self) -> i16 {
         let x_1st_to_8th_bit = if self.x_overflow() {
@@ -733,7 +844,7 @@ impl MousePacketGeneric {
     }
 
     /// `y_1st_to_8th_bit` and `y_9th_bit` should not be accessed directly, because they're part of one signed 9-bit number
-    /// 
+    ///
     /// Note: This handles overflow as well, although it might not happen on newer systems <https://forum.osdev.org/viewtopic.php?f=1&t=31176>
     pub fn y_movement(&self) -> i16 {
         let y_1st_to_8th_bit = if self.y_overflow() {
@@ -873,7 +984,7 @@ impl MousePacket {
 }
 
 #[derive(Clone)]
-pub enum SampleRate {
+pub enum MouseSampleRate {
     _10 = 10,
     _20 = 20,
     _40 = 40,
