--- kernel/ps2_controller/Cargo.toml
@@ -1,24 +1,0 @@
-[package]
-name = "ps2_controller"
-version = "0.1.0"
-description = "Driver for the 8042 PS/2 controller"
-authors = ["Hecatia Elegua"]
-edition = "2021"
-
-[dependencies]
-spin = "0.9.4"
-
-[dependencies.ps2]
-path = "../ps2"
-
-[dependencies.acpi]
-path = "../acpi"
-
-[dependencies.fadt]
-path = "../acpi/fadt"
-
-[dependencies.log]
-version = "0.4.8"
-
-[lib]
-crate-type = ["rlib"]
